requirements.txt
boto3==1.24.82
botocore==1.27.82
Cerberus==1.3.4
Command==0.1.0
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests_mock==1.10.0
coverage==6.5.0
cblack==22.6.0
pre-commit==2.20.0
